requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
requests==2.32.3
python-dotenv==1.0.1
websockets==16.0
google-generativeai==0.8.5
